add genYAML.py; clean up run_jedi_exe.py (#223)
